do dokonczenia

+++ MPR_spring/src/main/java/pl/edu/pjatk/MPR_spring/Service/CapybaraService.java
@@ -3,6 +3,8 @@ package pl.edu.pjatk.MPR_spring.Service;
 import org.apache.catalina.valves.rewrite.InternalRewriteMap;
 
 import org.springframework.stereotype.Component;
+import pl.edu.pjatk.MPR_spring.exception.CapybaraAlredyExist;
+import pl.edu.pjatk.MPR_spring.exception.CapybaraNotFoundException;
 import pl.edu.pjatk.MPR_spring.model.Capybara;
 import pl.edu.pjatk.MPR_spring.repository.CapybaraRepository;
 
@@ -24,45 +26,61 @@ public class CapybaraService {
         this.repository.save(new Capybara("andrey", "black"));
     }
     public List<Capybara> getCapybaraByName(String name) {
+        Iterable<Capybara> capybaraByNAme = repository.findAll();
+        capybaraByNAme.forEach(capybara -> stringUtilsService.lowerCase(String.valueOf(capybara)));
         return repository.findByName(name);
     }
     public List<Capybara> getCapybaraByColor(String color) {
+        Iterable<Capybara> capybaraByColor = repository.findAll();
+        capybaraByColor.forEach(capybara -> stringUtilsService.lowerCase(String.valueOf(capybara)));
         return repository.findByColor(color);
     }
 
 
 
     public Iterable<Capybara> getCapybaraList() {
+       Iterable<Capybara> capybaraIterable = repository.findAll();
+       capybaraIterable.forEach(capybara -> stringUtilsService.UpperCase(String.valueOf(capybara)));
+        return repository.findAll();
+    }
 
-        Iterable<Capybara> all = repository.findAll();
-        for (Capybara capybara : all) {
-
+    public Capybara getCapybara(Long id) {
+        Optional<Capybara> capybara = this.repository.findById(id);
+        if (capybara.isEmpty()) {
+            throw new CapybaraNotFoundException();
         }
-        return all;
-    }
-
-    public Capybara saveCapybara(Capybara capybara) {
-
-        capybara.setName(this.stringUtilsService.UpperCase(capybara.getName()));
-        capybara.setColor(this.stringUtilsService.UpperCase(capybara.getColor()));
-        return repository.save(capybara);
-    }
-
-    public Optional<Capybara> getCapybara(Long id) {
-        return this.repository.findById(id);
+        return capybara.get();
     }
 
     public void delete(Integer id) {
+        if(!repository.existsById(id.longValue())) {
+            throw new CapybaraNotFoundException();
+        }
         this.repository.deleteById(id);
+    }
+
+    public void add(Capybara capybara) {
+        if(repository.existsById(capybara.getId())) {
+            throw new CapybaraAlredyExist();
+        }
+        stringUtilsService.UpperCase(String.valueOf(capybara));
+        this.repository.save(capybara);
     }
 
 
     public void update(String name, String color, Capybara newCapybara) {
-        repository.findByName(name).stream().filter
-                (c -> c.getColor().equals(color)).forEach(existing -> { existing.setName(newCapybara.getName());
-            existing.setColor(newCapybara.getColor()); repository.save(existing);});
+        List<Capybara> capybaras = repository.findByName(name);
+        if (capybaras.isEmpty()) {
+            throw new CapybaraNotFoundException();
+        }else if(capybaras.get(0).getColor().equals(color)) {
+            throw new CapybaraAlredyExist();
+        }
+        capybaras.stream().filter
+                (c -> c.getColor().equals(color)).forEach(existing -> {
+                    existing.setName(newCapybara.getName());
+                    existing.setColor(newCapybara.getColor());
+                    repository.save(existing);});
     }
-
 
 
 }
